LPD-48254 SF - Move thread local to folder package
