+++ modules/apps/object/object-api/bnd.bnd
@@ -20,6 +20,7 @@ Export-Package:\
 	com.liferay.object.entries.frontend.data.set.data.model,\
 	com.liferay.object.entry,\
 	com.liferay.object.entry.contributor,\
+	com.liferay.object.entry.folder.util,\
 	com.liferay.object.entry.util,\
 	com.liferay.object.exception,\
 	com.liferay.object.field.attachment,\
